Moved Building Entities from BusinessLayer to BusinessEntities.Buildings
