+++ BusinessEntities/Buildings/Artillery.cs
@@ -1,11 +1,10 @@
-﻿using BusinessEntities;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BusinessLayer
+namespace BusinessEntities
 {
     public class Artillery : Building
     {
