+++ BusinessEntities/Buildings/CommandCenter.cs
@@ -3,9 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using BusinessEntities;
 
-namespace BusinessLayer
+namespace BusinessEntities
 {
     public class CommandCenter : Building
     {
