+++ BusinessEntities/Buildings/Mine.cs
@@ -3,18 +3,19 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using BusinessEntities;
 
-namespace BusinessLayer
+namespace BusinessEntities
 {
-    public class SatCom : Building
+    public class Mine : Building
     {
         //stores the value of the number of current soldiers in the building
         int soliersManned { get; set; }
         //stores the value of the number of maximum soldiers in the building
         int maxSoldiersManned { get; set; }
-        //stores the number of blocks that can be revealed.
-        int numBlocksRevealed { get; set; }
+        //base number of resources a player earns per turn for a mine
+        int resources { get; set; }
+        //modifier for resources earned based on the number of gold in adjacent terrain
+        int resourceModifier { get; set; }
