+++ BusinessEntities/Buildings/SatCom.cs
@@ -3,22 +3,16 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using BusinessEntities;
 
-namespace BusinessLayer
+namespace BusinessEntities
 {
-    public class Mine : Building
+    public class SatCom : Building
     {
         //stores the value of the number of current soldiers in the building
         int soliersManned { get; set; }
         //stores the value of the number of maximum soldiers in the building
         int maxSoldiersManned { get; set; }
-        //base number of resources a player earns per turn for a mine
-        int resources { get; set; }
-        //modifier for resources earned based on the number of gold in adjacent terrain
-        int resourceModifier { get; set; }
-
-
-
+        //stores the number of blocks that can be revealed.
+        int numBlocksRevealed { get; set; }
     }
 }
